+++ src/index.ts
@@ -3,6 +3,7 @@ import 'source-map-support/register';
 import * as path from 'path';
 import * as child_process from 'child_process';
 import * as commander from 'commander';
+import * as semver from 'semver';
 
 let homebridge;
 
@@ -30,7 +31,7 @@ class HomebridgeConfigUi {
       .option('-I, --insecure', '', () => config.homebridgeInsecure = true)
       .parse(process.argv);
 
-    if (process.env.HOMEBRIDGE_CONFIG_UI === '1' && process.env.CONFIG_UI_INSTALLED) {
+    if (process.env.HOMEBRIDGE_CONFIG_UI === '1' && semver.satisfies(process.env.CONFIG_UI_VERSION, '>=3.5.5')) {
       this.log(`Running in Docker Standalone Mode.`);
     } else if (config.noFork) {
       this.noFork(setup);